+++ gui/__init__.py
@@ -1,5 +1,5 @@
 from .util.misc import imshow
-from .elements import Window, Viewer
+from .components import Window, Viewer
 
 from ._version import get_versions
 __version__ = get_versions()['version']
