+++ gui/components/_control_bars/model.py
