+++ gui/components/_control_bars/view.py
