+++ gui/components/_window/view.py
