+++ gui/layers/_base_layer/__init__.py
@@ -1,2 +1,2 @@
 from .model import Layer
-from .view_props import QtLayer
+from .view import QtLayer
