+++ gui/layers/_markers_layer/view.py
